API

Domain Interface

ClimaAtmos.Domains.AbstractDomain — Type

AbstractDomain

ClimaAtmos.Domains.AbstractHorizontalDomain — Type

AbstractHorizontalDomain

ClimaAtmos.Domains.AbstractHybridDomain — Type

AbstractHybridDomain

ClimaAtmos.Domains.AbstractVerticalDomain — Type

AbstractVerticalDomain

ClimaAtmos.Domains.Column — Type

struct Column{FT} <: AbstractVerticalDomain

A column domain with zlim (Ordered Tuple) the domain extents, and nelements (Integer) the number of cells in the domain.

ClimaAtmos.Domains.Column — Type

Column([FT=Float64]; zlim, nelements)

Creates a column domain of type FT, with extents zlim[1] < zlim[2] and nelements cells.

Example: Generate a Column{Float64} with extents (0,1) and 10 elements.

julia> using ClimaAtmos.Domains
julia> z_domain = Column(Float64, 
                            zlim = (0,1), 
                            nelements = 10)

ClimaAtmos.Domains.HybridPlane — Type

struct HybridPlane{FT} <: AbstractHybridDomain

ClimaAtmos.Domains.HybridPlane — Type

HybridPlane([FT=Float64]; xlim, zlim, nelements, npolynomial)

Creates an 1D horizontal and 1D vertical hybrid domain with xlim the horizontal domain extents, zlim the vertical domain extents, helement the number of elements in the horizontal, velement the number of cells in the vertical, and npolynomial the polynomial order in the horizontal. ```

ClimaAtmos.Domains.Plane — Type

Plane <: AbstractHorizontalDomain

A two-dimensional specialisation of an AbstractHorizontalDomain. An x-y plane with extents xlim (Ordered Tuple) in the x-direction, extents ylim (Ordered Tuple) in the y-direction, nelements cells (Tuple of integers for x, y) directions, order of polynomial for spectral discretisation npolynomial, and description of periodicity in x, y directions periodic (Tuple of Booleans).

ClimaAtmos.Domains.Plane — Type

Plane([FT=Float64];
      xlim, 
      ylim, 
      nelements,
      npolynomial,
      periodic)

Creates an xy plane bounded by keyword argument values of xlim, ylim, with nelements elements (which may differ in each direction) of polynomial order npolynomial, and periodicity specified by periodic

NOTE: This definition of Plane currently supports Spectral Discretizations only. Updates will contain support for spectral-element (SpectralPlane) and spectral-element + finite difference (HybridPlane) configurations.

Example: Generate a Plane{Float64} object with extents [0,π] × [0,π], contains 5 and 10 elements in the x and y directions respectively, with polynomial order 5, and is non-periodic in the y-direction

julia> using ClimaAtmos.Domains
julia> xy_plane = Plane(Float64, xlim = (0,π), ylim = (0,π) , nelements=(5,10), npolynomial=5, periodic = (true,false))

ClimaAtmos.Domains.PeriodicPlane — Function

PeriodicPlane([FT= Float64];
    xlim,
    ylim,
    nelements,
    npolynomial)

Creates an xy plane bounded by keyword argument values of xlim, ylim, with nelements elements (which may differ in each direction) of polynomial order npolynomial. Assumes domain is periodic in both x and y directions. Special case of Plane.

NOTE: This definition of Plane currently supports Spectral Discretizations only. Updates will contain support for spectral-element (SpectralPlane) and spectral-element + finite difference (HybridPlane) configurations.

Example: Generate a Plane{Float64} object with extents [0,π] × [0,π], contains 5 and 10 elements in the x and y directions respectively, with polynomial order 5, and is doubly-periodic

julia> using ClimaAtmos.Domains
julia> xy_periodic_plane = PeriodicPlane(Float64, xlim = (0,π), ylim = (0,π) , nelements=(5,10), npolynomial=5)
# This is the same as 
julia> xy_periodic_plane_2 = Plane(Float64, xlim = (0,π), ylim = (0,π) , nelements=(5,10), npolynomial=5, periodic=(true,true))

ClimaAtmos.Domains.make_function_space — Method

make_function_space(domain::Column)

ClimaAtmos.Domains.make_function_space — Method

make_function_space(domain::HybridPlane)

ClimaAtmos.Domains.make_function_space — Method

make_function_space(domain::Plane)

Model Interface

ClimaAtmos.Models.AbstractModel — Type

AbstractModel

ClimaAtmos.Models.default_initial_conditions — Method

default_initial_conditions(model)

Construct the initial conditions for model.

ClimaAtmos.Models.get_boundary_flux — Function

get_boundary_flux

Construct the boundary fluxes.

ClimaAtmos.Models.make_ode_function — Method

make_ode_function(model)

Construct the ordinary differential equations for model.

Shallow Water Models

ClimaAtmos.Models.ShallowWaterModels.ShallowWaterModel — Type

struct ShallowWaterModel <: AbstractModel

Construct a shallow water model on domain with parameters.

Single Column Models

ClimaAtmos.Models.SingleColumnModels.SingleColumnModel — Type

SingleColumnModel <: AbstractModel

Construct a single column model on domain with parameters. boundary_conditions contains field boundary conditions.

Simulation Interface

ClimaAtmos.Simulations.AbstractSimulation — Type

AbstractSimulation

ClimaAtmos.Simulations.Simulation — Type

struct Simulation <: AbstractSimulation
    # a ClimaAtmos model
    model::AbstractModel
    # a DiffEqBase.jl integrator used for time
    # stepping the simulation
    integrator::DiffEqBase.DEIntegrator
    # user defined callback operations 
    callbacks::Union{DiffEqBase.CallbackSet, DiffEqBase.DiscreteCallback, Nothing}
end

A simulation wraps an abstract ClimaAtmos model containing equation specifications and an instance of an integrator used for time integration of the discretized model PDE.

ClimaAtmos.Simulations.Simulation — Method

Simulation(model::AbstractModel, method::AbstractTimestepper, 
           dt, tspan, init_state, callbacks, kwargs...)

Construct a Simulation for a model with a time stepping method, initial conditions Y_init, time step Δt for tspan time interval.

ClimaAtmos.Simulations.run! — Method

run!(simulation::AbstractSimulation, args...; kwargs...)

Run a simulation to the end.

ClimaAtmos.Simulations.set! — Function

set!(
    simulation::AbstractSimulation,
    submodel_name = nothing;
    kwargs...,
)

Set the simulation state to a new state, either through an array or a function.

SciMLBase.step! — Method

step!(simulation::AbstractSimulation, args...; kwargs...)

Step forward a simulation one time step.