API

ClimaAtmos.TurbulenceConvection.Parameters — Module

Parameters

Initial conditions

General

ClimaAtmos.InitialConditions.InitialCondition — Type

InitialCondition

A mechanism for specifying the LocalState of an AtmosModel at every point in the domain. Given some initial_condition, calling initial_condition(params) returns a function of the form local_state(local_geometry)::LocalState.

ClimaAtmos.InitialConditions.IsothermalProfile — Type

IsothermalProfile(; temperature = 300)

An InitialCondition with a uniform temperature profile.

ClimaAtmos.InitialConditions.DecayingProfile — Type

DecayingProfile(; perturb = true)

An InitialCondition with a decaying temperature profile, and with an optional perturbation to the temperature.

ClimaAtmos.InitialConditions.hydrostatic_pressure_profile — Function

hydrostatic_pressure_profile(; thermo_params, p_0, [T, θ, q_tot, z_max])

Solves the initial value problem p'(z) = -g * ρ(z) for all z ∈ [0, z_max], given p(0), either T(z) or θ(z), and optionally also q_tot(z). If q_tot(z) is not given, it is assumed to be 0. If z_max is not given, it is assumed to be 30 km. Note that z_max should be the maximum elevation to which the specified profiles T(z), θ(z), and/or q_tot(z) are valid.

Plane / Box

ClimaAtmos.InitialConditions.AgnesiHProfile — Type

AgnesiHProfile(; perturb = false)

An InitialCondition with a decaying temperature profile

ClimaAtmos.InitialConditions.ScharProfile — Type

ScharProfile(; perturb = false)

An InitialCondition with a prescribed Brunt-Vaisala Frequency

ClimaAtmos.InitialConditions.DryDensityCurrentProfile — Type

DryDensityCurrentProfile(; perturb = false)

An InitialCondition with an isothermal background profile, with a negatively buoyant bubble, and with an optional perturbation to the temperature.

ClimaAtmos.InitialConditions.RisingThermalBubbleProfile — Type

RisingThermalBubbleProfile(; perturb = false)

An InitialCondition with an isothermal background profile, with a positively buoyant bubble, and with an optional perturbation to the temperature.

Sphere

ClimaAtmos.InitialConditions.DryBaroclinicWave — Type

DryBaroclinicWave(; perturb = true)

An InitialCondition with a dry baroclinic wave, and with an optional perturbation to the horizontal velocity.

ClimaAtmos.InitialConditions.MoistBaroclinicWaveWithEDMF — Type

MoistBaroclinicWaveWithEDMF(; perturb = true)

The same InitialCondition as MoistBaroclinicWave, except with an initial TKE of 0 and an initial draft area fraction of 0.2.

ClimaAtmos.InitialConditions.MoistAdiabaticProfileEDMFX — Type

MoistAdiabaticProfileEDMFX(; perturb = true)

An InitialCondition with a moist adiabatic temperature profile, and with an optional perturbation to the temperature.

Cases from literature

ClimaAtmos.InitialConditions.Nieuwstadt — Type

Nieuwstadt

The InitialCondition described in [4], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.GABLS — Type

GABLS

The InitialCondition described in Kosovic2000, but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.GATE_III — Type

GATE_III

The InitialCondition described in [5], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.ARM_SGP — Type

ARM_SGP

The InitialCondition described in [6], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.DYCOMS_RF01 — Type

DYCOMS_RF01

The InitialCondition described in [7], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.DYCOMS_RF02 — Type

DYCOMS_RF02

The InitialCondition described in [8], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.Rico — Type

Rico

The InitialCondition described in Rauber2007, but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.TRMM_LBA — Type

TRMM_LBA

The InitialCondition described in [9], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.LifeCycleTan2018 — Type

LifeCycleTan2018

The InitialCondition described in [10], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.Bomex — Type

Bomex

The InitialCondition described in [11], but with a hydrostatically balanced pressure profile.

ClimaAtmos.InitialConditions.Soares — Type

Soares

The InitialCondition described in [12], but with a hydrostatically balanced pressure profile.