API

Domains

ClimaAtmos.Domains.AbstractDomain — Type

Supertype for all domains.

ClimaAtmos.Domains.Column — Method

Column([FT = Float64]; zlim, nelements, stretching)

Construct a domain of type FT that represents a column along the z-axis with limits zlim (where zlim[1] < zlim[2]) and nelements elements. This domain is not periodic.

Example

julia> Column(zlim = (0, 1), nelements = 10)
Domain set-up:
	z-column:	[0.0, 1.0]
	# of elements:	10
	Vertical stretching rule:	Uniform()

ClimaAtmos.Domains.HybridBox — Method

HybridBox([FT = Float64]; xlim, ylim, zlim, nelements, npolynomial, xperiodic = true, yperiodic = true, stretching

Construct a domain of type FT that represents an xz-plane with limits xlim ylim and zlim (where xlim[1] < xlim[2],ylim[1] < ylim[2], and zlim[1] < zlim[2]), nelements elements of polynomial order npolynomial, x-axis periodicity xperiodic, and y-axis periodicity yperiodic. nelements must be a tuple with two values, with the first value corresponding to the x-axis, the second corresponding to the y-axis, and the third corresponding to the z-axis. This domain is not periodic along the z-axis. If stretching is unspecified, generates a uniform vertical interval mesh.

Example

julia> HybridBox(
            xlim = (0, π),
            ylim = (0, π),
            zlim = (0, 1),
            nelements = (5, 5, 10),
            npolynomial = 5,
            xperiodic = true,
            yperiodic = true,
            )
Domain set-up:
	xyz-box:	[0.0, 3.1) × [0.0, 3.1) × [0.0, 1.0]
	# of elements:	(5, 5, 10)
	poly order:	5
	Vertical stretching rule:	Uniform()

ClimaAtmos.Domains.HybridPlane — Method

HybridPlane([FT = Float64]; xlim, zlim, nelements, npolynomial, xperiodic = true, stretching)

Construct a domain of type FT that represents an xz-plane with limits xlim and zlim (where xlim[1] < xlim[2] and zlim[1] < zlim[2]), nelements elements of polynomial order npolynomial, and x-axis periodicity xperiodic. nelements must be a tuple with two values, with the first value corresponding to the x-axis and the second corresponding to the z-axis. This domain is not periodic along the z-axis. If the stretching keyword is unspecified, a uniform vertical mesh is generated.

Example

julia> HybridPlane(
            xlim = (0, π),
            zlim = (0, 1),
            nelements = (5, 10),
            npolynomial = 5,
            xperiodic = true,
        )
Domain set-up:
	xz-plane:	[0.0, 3.1) × [0.0, 1.0]
	# of elements:	(5, 10)
	poly order:	5
	Vertical stretching rule:	Uniform()

ClimaAtmos.Domains.SphericalShell — Method

SphericalShell([FT = Float64]; radius, height, nelements, npolynomial, stretching)

Construct a domain of type FT that represents a spherical shell with radius radius, height height, and nelements elements of polynomial order npolynomial. If stretching is unspecified, generates a uniform vertical interval mesh.

Example

julia> SphericalShell(radius = 1, height = 1, nelements = (6, 10), npolynomial = 5)
Domain set-up:
	sphere radius:	1.0
	sphere height:	1.0
	# of elements:	(6, 10)
	poly order:	5
	Vertical stretching rule:	Uniform()

ClimaAtmos.Domains.make_function_space — Function

make_function_space(domain)

Convert an AbstractDomain into a ClimaCore.Spaces.AbstactSpace.

Boundary Conditions

ClimaAtmos.BoundaryConditions.AbstractBoundary — Type

Supertype for all boundary conditions.

ClimaAtmos.BoundaryConditions.BulkFormula — Type

BulkFormula{C, T} <: AbstractBoundary

Computes the boundary flux using the bulk formula and constant or Mohnin-Obukhov-based heat transfer coefficient Ch.

ClimaAtmos.BoundaryConditions.DragLaw — Type

DragLaw{C} <: AbstractBoundary

Computes the boundary flux using the bulk formula and constant or Mohnin-Obukhov-based drag coefficient Cd. Specific to momentum density.

ClimaAtmos.BoundaryConditions.NoFlux — Type

NoFlux <: AbstractBoundary

Computes a fixed boundary flux of 0.

ClimaAtmos.BoundaryConditions.NoVectorFlux — Type

NoVectorFlux <: AbstractBoundary

Computes a fixed boundary vector flux of 0

ClimaAtmos.BoundaryConditions.get_boundary_flux — Function

get_boundary_flux(model, bc, var, Ym, Ya)

Get the flux of variable var across the boundary of model using boundary condition bc.

ClimaAtmos.BoundaryConditions.get_boundary_flux — Method

get_boundary_flux(model, ::NoFlux, var::Fields.Field, Y, Ya)

Vertical flux assignment for no-flux conditions (e.g., insulating wall)

ClimaAtmos.BoundaryConditions.get_boundary_flux — Method

get_boundary_flux(model, ::NoVectorFlux, var::Fields.Field, Y, Ya)

Vertical flux assignment for an AxisTensor with CovariantAxis{(1,2)} in no-flux conditions (e.g., free-slip wall)

Models Interface

ClimaAtmos.Models.AbstractBaseModelStyle — Type

Supertype for all base model styles.

ClimaAtmos.Models.AbstractCacheStyle — Type

Supertype for all model cache styles.

ClimaAtmos.Models.AbstractModel — Type

Supertypes for all models.

ClimaAtmos.Models.AbstractModelStyle — Type

Supertype for all model styles.

ClimaAtmos.Models.AbstractMoistureStyle — Type

Supertype for all model moisture styles.

ClimaAtmos.Models.AbstractPrecipitationStyle — Type

Supertype for all model precipitation styles.

ClimaAtmos.Models.AbstractThermodynamicsStyle — Type

Supertype for all model thermodynamics styles.

ClimaAtmos.Models.AbstractVerticalDiffusionStyle — Type

Supertype for all vertical diffusion styles.

ClimaAtmos.Models.components — Method

components(model::AbstractModel)

Return the components of model, e.g. base, thermodynamics, moisture, etc.

ClimaAtmos.Models.default_initial_conditions — Method

default_initial_conditions(model, space_center, space_face)

Construct the initial conditions for model.

ClimaAtmos.Models.default_ode_cache — Method

default_ode_cache(model, cache, space_center, space_face)

Construct the cache for ordinary differential equations.

ClimaAtmos.Models.get_velocities — Method

get_velocities(Y, model)

Construct the initial conditions for model.

ClimaAtmos.Models.make_ode_function — Method

make_ode_function(model)

Construct the ordinary differential equations for model.

ClimaAtmos.Models.variable_names — Method

variables_names(styles::AbstractModelStyle)

Return the state variable names for style.

Example

julia> Models.variable_names(NonEquilibriumMoisture())
(:ρq_tot, :ρq_liq, :ρq_ice)

ClimaAtmos.Models.variable_spaces — Method

variables_spaces(style::AbstractModelStyle, model::AbstractModel)

Return the state variables for style.

ClimaAtmos.Models.variable_types — Method

variables_types(style::AbstractModelStyle, model::AbstractModel, FT)

Return the state variable types for style and model with float type FT.

Models

ClimaAtmos.Models.SingleColumnModels.SingleColumnModel — Type

SingleColumnModel <: AbstractModel

A single column model. Required fields are domain, boundary_conditions, and parameters.

ClimaAtmos.Models.Nonhydrostatic2DModels.Nonhydrostatic2DModel — Type

Nonhydrostatic2DModel <: AbstractModel

A two-dimensional non-hydrostatic model, which is typically used for simulating the Euler equations. Required fields are domain, boundary_conditions, and parameters.

ClimaAtmos.Models.Nonhydrostatic3DModels.Nonhydrostatic3DModel — Type

Nonhydrostatic3DModel <: AbstractModel

A three-dimensional non-hydrostatic model, which is typically used for simulating the Euler equations. Required fields are domain, boundary_conditions, and parameters.

Callbacks

ClimaAtmos.Callbacks.AbstractCallback — Type

Supertype for all callbacks.

ClimaAtmos.Callbacks.CFLAdaptive — Type

CFLAdaptive <: AbstractCallback

Container for CFL information callback.

ClimaAtmos.Callbacks.JLD2Output — Type

JLD2Output{M, I} <: AbstractCallback

Specifies that a DiffEqCallbacks.PeriodicCallback should be constructed that extracts the model state from the integrator and stores it in a .jld2 file.

ClimaAtmos.Callbacks.domain_max_cfl — Method

(F::CFLAdaptive)(integrator)

For the given model and function space, gets the nodal distances (local=> horizontal), (minimum=> vertical) and model velocities from which the CFL number is calculated at the user specified time interval. If the update flag is true, sets new timestep size.

ClimaAtmos.Callbacks.generate_callback — Function

generate_callback(callback; kwargs...)

Convert an AbstractCallback to a SciMLBase.DECallback.

ClimaAtmos.Callbacks.generate_callback — Method

generate_callback(F::CFLAdaptive; kwargs...)

Creates a PeriodicCallback object that computes the
maximum CFL number in the domain.

ClimaAtmos.Callbacks.get_nodal_distance — Method

get_nodal_distance(space::Space)

Return a tuple of local spacing for a ClimaCore Space object. Dispatched over AbstractSpace subtypes.

Simulations

ClimaAtmos.Simulations.AbstractRestart — Type

Supertype for all restart modes.

ClimaAtmos.Simulations.NoRestart — Type

NoRestart <: AbstractRestart

Specifies that a simulation should use its original initial conditions and end time.

ClimaAtmos.Simulations.Restart — Type

Restart <: AbstractRestart

Specifies that a simulation should begin from the state recorded in a restart file and end at a specific time. The restart file must be a .jld2 file containing the simulation.integrator and simulation.model objects. Users must set! the simulation prior to restart.

ClimaAtmos.Simulations.Simulation — Method

Simulation(
    model,
    method;
    Y_init = nothing,
    dt,
    tspan,
    callbacks = nothing,
    restart = NoRestart(),
)

Construct a Simulation for a model with a time stepping method, initial conditions Y_init, and time step dt for a time interval of tspan. If Y_init is not provided, the model's default initial conditions are used. ```

ClimaAtmos.Simulations.get_spaces — Function

get_spaces(simulation::Simulation, subcomponent = :base; args...)

Extract the function space for each field in args within model subcomponents subcomponent.

ClimaAtmos.Simulations.run! — Method

run!(simulation, args...; kwargs...)

Run the simulation to completion.

ClimaAtmos.Simulations.set! — Function

set!(simulation, subcomponent = :base; kwargs...)

Set the simulation state to a new state, either through an array or a function. Defaults to :base component for model.

ClimaAtmos.Simulations.step! — Method

step!(simulation, args...; kwargs...)

Advance the simulation by one time step.

Initial Conditions

ClimaAtmos.Utils.InitialConditions.init_1d_ekman_column — Method

init_1d_ekman_column(params)

Ekman layer initial condition for 1D single-column benchmarking.
Reference parameter values:
    - T_surf = 300 K
    - T_min_ref = 230 K
    - u0 = 1 ms⁻¹
    - v0 = 0 ms⁻¹
    - w0 = 0 ms⁻¹
    - θ_b = 300 K
    - p_0 = 1e5 Pa
    - g = 9.80616 m s⁻²
    - R_d = 287.0 J kg⁻¹ K⁻¹
    - cp_d = 1004 J kg⁻¹ K⁻¹

ClimaAtmos.Utils.InitialConditions.init_2d_dry_bubble — Method

init_2d_dry_bubble(params, thermovar = :ρθ)

Rising bubble initial condition for 2D box benchmarking.
Reference: https://journals.ametsoc.org/view/journals/mwre/140/4/mwr-d-10-05073.1.xml, Section 5a
Reference parameter values:
    - x_c = 0 m
    - z_c = 350 m
    - r_c = 250 m
    - θ_b = 300 K
    - θ_c = 0.5 K
    - p_0 = 1e5 Pa
    - cp_d = 1004 J kg⁻¹ K⁻¹
    - cv_d = 717.5 J kg⁻¹ K⁻¹
    - R_d = 287.0 J kg⁻¹ K⁻¹
    - g = 9.80616 m s⁻²

ClimaAtmos.Utils.InitialConditions.init_2d_moist_bubble — Method

init_2d_dry_bubble(params, thermovar = :ρθ)

Rising bubble initial condition for 2D box benchmarking.
    Reference: https://journals.ametsoc.org/view/journals/mwre/140/4/mwr-d-10-05073.1.xml, Section 5a    Reference parameter values:
    - x_c = 0 m
    - z_c = 350 m
    - r_c = 250 m
    - θ_b = 300 K
    - θ_c = 0.5 K
    - p_0 = 1e5 Pa
    - g = 9.80616 m s⁻²

ClimaAtmos.Utils.InitialConditions.init_2d_precipitating_bubble — Method

init_2d_dry_bubble(params, thermovar = :ρθ)

Rising bubble initial condition for 2D box benchmarking.
    Reference: https://journals.ametsoc.org/view/journals/mwre/140/4/mwr-d-10-05073.1.xml, Section 5a    Reference parameter values:
    - x_c = 0 m
    - z_c = 350 m
    - r_c = 250 m
    - θ_b = 300 K
    - θ_c = 0.5 K
    - p_0 = 1e5 Pa
    - g = 9.80616 m s⁻²

ClimaAtmos.Utils.InitialConditions.init_3d_baroclinic_wave — Method

init_3d_baroclinic_wave(FT, params, isbalanced = true)

- isbalanced: false is the standard baroclinic set up as in the reference;
              true is the same background flow without perturbation.
Dry shallow atmosphere baroclinic wave initial condition for 3D sphere benchmarking.
Reference: https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/qj.2241

ClimaAtmos.Utils.InitialConditions.init_3d_solid_body_rotation — Method

init_solid_body_rotation(FT, params)

Hydrostatically balanced initial condition for 3D sphere benchmarking.