List of functions in ClimaAtmos.jl module

ClimaAtmos.function_documentation_template — Method

functiondocumentationtemplate(a; informative_keyword = "yes")

Description

The goal of this function is to provide a template for all functions used in ClimaAtmos.jl

Arguments

a | Type: Any | An example function argument

Keyword Arguments

informative_keyword. Default: "yes"

Return

π : a number

ClimaAtmos.@boilerplate — Macro

macro boilerplate()

Description

A convenience function that includes all the usual needed for a script